shift compare add int tests to release
